backend/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
firebase-admin==6.6.0
pydantic==2.10.4
python-dotenv==1.0.1
google-genai>=1.56.0
google-cloud-speech>=2.27.0
google-cloud-storage>=2.18.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
fpdf2>=2.8.1
httpx>=0.27.0
cryptography>=42.0
